requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0

# Optional but recommended
gunicorn==21.2.0  # For production deployment